Added APA back to collaborations list
Put back in APA collaboration that was accidentally deleted during edits.
Note: this pull request includes all the edits from the other ones.   It would be easier it this point to reject the other pull requests and just accept this one.   The conflicts may also go away when this is done, but there may be some other ones that have to be resolved manually.