Update timeline
First cut at timeline based on a July 1 start and end date.  Please make suggested edits as appropriate.  

One tricky bit is how to interleave the "first" Profile publication and the "updated" Profile publication.  Because the short charter duration we'll still be waiting for the REC of the first Profile doc while we're well into the CR/PR process for the updated Profile doc.  It would be good however to merge these schedules at the end of the charter and do all the "updates" together.