Remove links to template
There were a bunch of links to internal sections that actually pointed to the template.  Fixed to be relative links.  Also removed a commented-out section that was not useful.

There is still one link to the template in the history table that we should update eventually when the charter is finalized.